fix(personal-context): address parent PR review
