+++ apps/tldw-frontend/lib/api/openapi.fingerprint.json
@@ -3,5 +3,5 @@
   "openapi_version": "3.1.0",
   "path_count": 2064,
   "schema_count": 3111,
-  "sha256": "30194b62b984168e8897826a77e173eff7cc5a6ed922f4a818db06b44ef72f76"
+  "sha256": "f330ccdff9bd24e4078e12443af6e5576946df76893f5617f98d0a078f3eaddf"
 }
